Service class 통합
