+++ src/main/java/Hairyou/demo/designer/service/DesignerAuthService.java
@@ -1,17 +1,22 @@
 package Hairyou.demo.designer.service;
 
+import Hairyou.demo.designer.repository.dto.DesignerDto;
 import Hairyou.demo.designer.repository.entity.Designer;
 import Hairyou.demo.designer.repository.entity.DesignerRepository;
-import Hairyou.demo.designer.repository.dto.DesignerDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class designerSignupService {
-
+public class DesignerAuthService {
     private final DesignerRepository designerRepository;
-
+    public Designer login_designer(String id, String password) {
+        Designer designer = designerRepository.findByid(id);
+        if (designer != null && designer.getPassword().equals(password)) {
+            return designer;
+        }
+        return null;
+    }
     public String registerDesigner(DesignerDto designer) {
         // Implement validation or other business logic if needed
         designerRepository.save(Designer.builder().id(designer.getID()).password(designer.getPassword()).nickname(designer.getNickname()).career(designer.getCareer()).manager(designer.getManager()).designerName(designer.getDesignerName()).build());
