+++ src/main/java/Hairyou/demo/user/controller/UserAuthController.java
@@ -2,7 +2,7 @@ package Hairyou.demo.user.controller;
 
 import Hairyou.demo.user.repository.dto.UserDto;
 import Hairyou.demo.user.repository.entity.User;
-import Hairyou.demo.user.service.userLoginService;
+import Hairyou.demo.user.service.userAuthService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -12,18 +12,18 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 @RequestMapping("/api/auth/user")
 public class UserAuthController {
-    private final Hairyou.demo.user.service.userSignupService userSignupService;
-    private final userLoginService loginService;
+    private final userAuthService userAuthService;
+
     @PostMapping("/sign-up")
     public ResponseEntity registerUser(@RequestBody UserDto userDto) {
-        if(userSignupService.registerUser(userDto).equals("Success")){
+        if(userAuthService.registerUser(userDto).equals("Success")){
             return new ResponseEntity(HttpStatus.CREATED);
         };
         return new ResponseEntity(HttpStatus.BAD_REQUEST);
     }
     @PostMapping("/login")
     public ResponseEntity<String> user_login(@RequestParam String id , @RequestParam String password){
-        User user=loginService.login_user(id, password);
+        User user=userAuthService.login_user(id, password);
         if (user != null) {
             return ResponseEntity.ok("Login successful for user: " + user.getUserName());
         } else {
